- renamed generateRoot.js to merkleTreeService.js
