+++ frontend/src/scripts/merkleTreeService.js
